+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
-#### 11.3.0 - Unreleased
+#### 11.4.0 - Unreleased
 * Add Fable support to FsLexYacc.Runtime.
+* Make the AssocTable lookup cache initial capacity configurable to avoid pre-allocating 2000 entries per parse #54
+* Add `--assoc-cache-capacity` option for fsyacc to set the generated parser's AssocTable cache capacity from the command line #54
 
 #### 11.2.0 - 12 May, 2023
 * Add `--open` option for fslex.